+++ tests/functions/function_handle_from_name.m
@@ -1,0 +1,9 @@
+% Test: Function handle from named function
+% EXPECT: warnings = 0
+% EXPECT: f = function_handle
+
+function y = myFunc(x)
+    y = x + 1;
+end
+
+f = @myFunc;